hello-game-app/hello-function/requirements.txt
google-cloud-pubsub==2.33.0
cloud-sql-python-connector==1.18.5
pg8000==1.31.5